requirements.txt
tensorflow=2.10
opencv-python
Pillow
tk
